# Improve performance serialize to json at ntex