--- a/frameworks/Rust/ntex/ntex-db.dockerfile
+++ b/frameworks/Rust/ntex/ntex-db.dockerfile
@@ -1,5 +1,8 @@
 FROM rust:1.44
 
+# Disable simd at jsonescape
+ENV CARGO_CFG_JSONESCAPE_DISABLE_AUTO_SIMD=
+
 RUN apt-get update -yqq && apt-get install -yqq cmake g++
 
 ADD ./ /ntex