--- a/frameworks/Rust/ntex/src/utils.rs
+++ b/frameworks/Rust/ntex/src/utils.rs
@@ -1,22 +1,9 @@
 #![allow(dead_code)]
-use std::{cmp, io};
+use std::cmp;
 
 use atoi::FromRadix10;
-use bytes::BytesMut;
 
 pub const SIZE: usize = 27;
-
-pub struct Writer<'a>(pub &'a mut BytesMut);
-
-impl<'a> io::Write for Writer<'a> {
-    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-        self.0.extend_from_slice(buf);
-        Ok(buf.len())
-    }
-    fn flush(&mut self) -> io::Result<()> {
-        Ok(())
-    }
-}
 
 pub fn get_query_param(query: &str) -> u16 {
     let q = if let Some(pos) = query.find("q") {