@fafhrd91 I have forked `simd-json-derive` I have implemented it directly over the `bytes::BytesMut`. The advantage is that it does not use intermediate buffers, serializing on a single alloc. This is perfect for the `Word` struct, as seen in the result:
### Old
```json
 "db": {
      "ntex-db": [
        {
          "latencyAvg": "331.62us",
          "latencyMax": "11.84ms",
          "latencyStdev": "217.48us",
          "totalRequests": 732443,
          "startTime": 1593489086,
          "endTime": 1593489101
        },
        {
          "latencyAvg": "589.14us",
          "latencyMax": "10.34ms",
          "latencyStdev": "382.99us",
          "totalRequests": 837341,
          "startTime": 1593489103,
          "endTime": 1593489118
        },
        {
          "latencyAvg": "1.05ms",
          "latencyMax": "12.67ms",
          "latencyStdev": "738.02us",
          "totalRequests": 962237,
          "startTime": 1593489120,
          "endTime": 1593489135
        },
        {
          "latencyAvg": "1.91ms",
          "latencyMax": "16.50ms",
          "latencyStdev": "1.41ms",
          "totalRequests": 1061544,
          "startTime": 1593489137,
          "endTime": 1593489152
        },
        {
          "latencyAvg": "3.53ms",
          "latencyMax": "35.39ms",
          "latencyStdev": "2.47ms",
          "totalRequests": 1127646,
          "startTime": 1593489154,
          "endTime": 1593489169
        },
        {
          "latencyAvg": "6.54ms",
          "latencyMax": "47.44ms",
          "latencyStdev": "3.93ms",
          "totalRequests": 1179279,
          "startTime": 1593489171,
          "endTime": 1593489186
        }
      ]
    },
```

#### New
```json
"db": {
      "ntex-db": [
        {
          "latencyAvg": "325.77us",
          "latencyMax": "11.29ms",
          "latencyStdev": "238.57us",
          "totalRequests": 747967,
          "startTime": 1593488128,
          "endTime": 1593488143
        },
        {
          "latencyAvg": "568.11us",
          "latencyMax": "10.99ms",
          "latencyStdev": "331.87us",
          "totalRequests": 862393,
          "startTime": 1593488145,
          "endTime": 1593488160
        },
        {
          "latencyAvg": "1.02ms",
          "latencyMax": "12.46ms",
          "latencyStdev": "720.79us",
          "totalRequests": 984987,
          "startTime": 1593488162,
          "endTime": 1593488177
        },
        {
          "latencyAvg": "1.91ms",
          "latencyMax": "23.67ms",
          "latencyStdev": "1.49ms",
          "totalRequests": 1079151,
          "startTime": 1593488179,
          "endTime": 1593488194
        },
        {
          "latencyAvg": "3.38ms",
          "latencyMax": "27.17ms",
          "latencyStdev": "2.26ms",
          "totalRequests": 1165541,
          "startTime": 1593488196,
          "endTime": 1593488211
        },
        {
          "latencyAvg": "6.28ms",
          "latencyMax": "47.12ms",
          "latencyStdev": "3.61ms",
          "totalRequests": 1217962,
          "startTime": 1593488213,
          "endTime": 1593488228
        }
      ]
    },
```

However I have not used it in `Message` struct because `simd-json-derive` escape algorithm is a little faster in short cases, less than 16 bytes, than `v_jsonescape`