@botika 2020-06-30
I have used the scalar version of the json escape. The profit with the previous one is an `if`. The difference is nil. I have cleaned the useless parts.